requirements.txt
accelerate==1.8.1
altair==5.5.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
decorator==5.2.1
diffusers==0.34.0
dm-tree==0.1.9
easyocr==1.7.2
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-auth==2.40.3
google-auth-oauthlib==0.4.6
google-cloud-vision==3.10.2
google-pasta==0.2.0
googleapis-common-protos==1.70.0
googletrans==4.0.0rc1
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.9.0
h2==3.2.0
h5py==3.14.0
hf-xet==1.1.5
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.33.2
hyperframe==5.2.0
idna==2.10
imageio==2.37.0
importlib-metadata==6.11.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==2.11.0
langdetect==1.0.9
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.1.0
narwhals==1.45.0
networkx==3.4.2
ninja==1.11.1.4
numpy==1.26.4
oauthlib==3.3.1
onnx==1.12.0
onnx-tf==1.10.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optimum==1.26.1
optree==0.16.0
packaging==23.2
pandas==2.3.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pydeck==0.9.1
Pygments==2.19.2
Pympler==1.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.46.1
streamlit-drawable-canvas==0.8.0
sympy==1.14.0
tenacity==8.5.0
termcolor==3.1.0
tf_keras==2.19.0
tifffile==2025.5.10
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.53.0